pyproject.toml
[project]
name = "travel-planner-with-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.31.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
